+++ dgdg-api/build.gradle
@@ -3,8 +3,6 @@ ext {
 }
 
 dependencies {
-
-    implementation 'junit:junit:4.13.1'
     implementation project(":dgdg-domain")
     implementation project(":dgdg-common")
