+++ dgdg-api/src/main/java/com/depromeet/dgdg/external/kakao/dto/response/KakaoUserResponse.java
@@ -7,7 +7,6 @@ import lombok.*;
 @ToString
 @Getter
 @NoArgsConstructor
-@AllArgsConstructor
 public class KakaoUserResponse {
 
     private String id;
@@ -17,10 +16,10 @@ public class KakaoUserResponse {
     @ToString
     @Getter
     @NoArgsConstructor
-    @AllArgsConstructor
     @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-    public static class Properties {
+    private static class Properties {
         private String nickname;
+        private String thumbnailImage;
         private String profileImage;
     }
