+++ dgdg-api/src/main/java/com/depromeet/dgdg/service/auth/dto/request/AuthRequest.java
@@ -16,8 +16,4 @@ public class AuthRequest {
     @NotBlank(message = "accessToken을 입력해주세요.")
     private String accessToken;
 
-    public static AuthRequest testInstance(String accessToken) {
-        return new AuthRequest(accessToken);
-    }
-
 }
